refactor: extract release version target checks
